move simplefempp into simplefem.py
